pyproject.toml
[tool.poetry]
name = "protocol-compose"
version = "0.2.0"
description = "Describe field trial protocols inspired by Docker Compose"
authors = ["João Murilo Palone Fauvel <joaofauvel@fundacaomt.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
pydantic = "^2.7.1"
rich = "^13.7.1"
more-itertools = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
